fix：修复多类 bug + 统一异常处理链路
- 修复 Sect.deductFunds 布尔值反转、SectBuildingService/SectMemberService 条件判断同步修正
- 修复 DamageCalculator replace 换 replaceAll(\b) 避免变量名前缀误匹配，解析器 int→long 防溢出
- 修复 TeamService 成员计数改为原子 SQL UPDATE 消除竞态
- 修复 BeastMutationService 变异从随机改为从剩余可用 trait 中抽取
- 修复 PostCombatProcessor 战斗者 ID 匹配、TravelCompleter 途中死亡/属性阈值检查
- BuffManager 攻防速修正添加上限、DefaultCombatEngine AOE_DAMAGE/HEAL 公式修正
- AuthenticatedAspect: BusinessException 透传保证 @Transactional 回滚，新增 CommandHandlerHelper.safeCall() 统一异常转错误文本
- 17 个 CommandHandler 全部迁移到 safeCall() 消除重复的 switch 模式匹配
- 其他防御性修复: SpiritStoneService 参数校验、BeastProductionService 时钟偏移、PriceEngine 最小价、UserStateService 行锁、TribulationService 防并发

+++ src/main/java/top/stillmisty/xiantao/service/enhance/EnhancementService.java
@@ -117,7 +117,7 @@ public class EnhancementService {
     }
 
     int stoneCost = core.calculateSpiritStoneCost(equipment.getRarity(), targetLevel);
-    int balance = spiritStoneService.getBalance(userId);
+    long balance = spiritStoneService.getBalance(userId);
     if (balance < stoneCost) {
       throw new BusinessException(ErrorCode.SPIRIT_STONES_INSUFFICIENT, stoneCost, balance);
     }
